docs: mocking app component dependencies in test
